requirements-app.txt
pyinstaller>=6.0.0
pystray>=0.19.5
pillow>=10.0.0
pywebview>=5.0
